requirements.txt
matplotlib >= 1.4.0
ImageHash
Pillow >= 3.0.0
